requirements.txt
# CerTrack - Student Certification & Event Platform
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
python-dotenv==1.0.0
email-validator==2.1.0
Werkzeug==3.0.1

# For PostgreSQL (optional - SQLite used by default):
# psycopg2-binary>=2.9
